+++ actix-macros/src/lib.rs
@@ -11,7 +11,7 @@ use quote::quote;
 ///
 /// ## Usage
 ///
-/// ```rust
+/// ```
 /// #[actix_rt::main]
 /// async fn main() {
 ///     println!("Hello world");