+++ actix-server/CHANGES.md
@@ -1,6 +1,9 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+* Hidden `ServerBuilder::start` method has been removed. Use `ServerBuilder::run`. [#246]
+
+[#246]: https://github.com/actix/actix-net/pull/246
 
 
 ## 2.0.0-beta.2 - 2021-01-03