+++ actix-server/CHANGES.md
@@ -2,9 +2,10 @@
 
 ## Unreleased - 2021-xx-xx
 * Hidden `ServerBuilder::start` method has been removed. Use `ServerBuilder::run`. [#246]
-* Add retry for EINTR(`io::Interrupted`) in `Accept`'s poll loop.
+* Add retry for EINTR(`io::Interrupted`) in `Accept`'s poll loop. [#264]
 
 [#246]: https://github.com/actix/actix-net/pull/246
+[#264]: https://github.com/actix/actix-net/pull/264
 
 
 ## 2.0.0-beta.2 - 2021-01-03